+++ graphbrain/corefs/onto.py
@@ -1,11 +1,11 @@
 from graphbrain.corefs import make_corefs
+from graphbrain.processor import Processor
 from graphbrain.utils.ontology import subtypes
 
 
-class CorefsOnto:
+class CorefsOnto(Processor):
     def __init__(self, hg, sequence=None):
-        self.hg = hg
-        self.sequence = sequence
+        super().__init__(hg=hg, sequence=sequence)
         self.corefs = 0
         self.done = set()
 
@@ -35,33 +35,25 @@ class CorefsOnto:
                 dd = self.hg.deep_degree(edge)
 
                 if dd > sdd:
-                    sdd_dd = float(sdd) / float(dd)
+                    # sdd_dd = float(sdd) / float(dd)
 
-                    self.logger.debug('concept: {}'.format(edge.to_str()))
-                    self.logger.debug('subconcepts: {}'.format(subs))
-                    self.logger.debug('# subs: {}'.format(len(subs)))
-                    self.logger.debug('max_ratio: {}'.format(max_ratio))
-                    self.logger.debug('sdd: {}'.format(sdd))
-                    self.logger.debug('dd: {}'.format(dd))
-                    self.logger.debug('sdd_dd: {}'.format(sdd_dd))
+                    # print('concept: {}'.format(edge.to_str()))
+                    # print('subconcepts: {}'.format(subs))
+                    # print('# subs: {}'.format(len(subs)))
+                    # print('max_ratio: {}'.format(max_ratio))
+                    # print('sdd: {}'.format(sdd))
+                    # print('dd: {}'.format(dd))
+                    # print('sdd_dd: {}'.format(sdd_dd))
 
                     if max_ratio >= .7:  # and sdd_dd < .5:
                         edge1 = edge
                         edge2 = subs[best_pos]
 
-                        self.logger.debug(
-                            'are corefs: {} | {}'.format(edge1.to_str(),
-                                                         edge2.to_str()))
+                        # print('are corefs: {} | {}'.format(edge1.to_str(),
+                        #                                    edge2.to_str()))
 
                         self.corefs += 1
                         make_corefs(self.hg, edge1, edge2)
 
-    def run(self):
-        print('processing edges...')
-        if self.sequence is None:
-            for edge in self.hg.all():
-                self.process_edge(edge)
-        else:
-            for edge in self.hg.sequence(self.sequence):
-                self.process_edge(edge)
-        print('{} coreferences were added.'.format(str(self.corefs)))
+    def report(self):
+        return '{} coreferences were added.'.format(str(self.corefs))