+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - New package graphbrain.meaning, for meaning extractors.
 - New package graphbrain.readers, for text readers from various sources.
 - New package graphbrain.corefs, for coreferences resolution.
+- Processor class.
 
 ### Changed
 - Sequences keep track of their size, no need to specify position when adding edge to sequence.